refactor: Separate discord and browser logic away from util.py
