+++ src/commands/card.py
@@ -3,6 +3,7 @@ from src.commands.base import BaseCommand
 from src.trie import Trie
 import src.util as util
 import src.curiosa as curiosa
+from src.discord import code_blockify
 
 
 class CardCommand(BaseCommand):
@@ -29,4 +30,4 @@ class CardCommand(BaseCommand):
             card_name, self.pt, self.cards
         )
 
-        return util.code_blockify(received_output)
+        return code_blockify(received_output)
